feature: review meta에 isOpen 필드 추가

+++ src/main/kotlin/jjfactory/organization/review/meta/ReviewMetaServiceImpl.kt
@@ -16,6 +16,7 @@ class ReviewMetaServiceImpl(
 
     override fun updateMeta(id: Long, request: ReviewMetaDto.UpdateRequest): Long {
         val meta = reviewMetaRepository.findByIdOrNull(id) ?: throw NotFoundException()
+        if (meta.isOpen) throw IllegalArgumentException("open 상태 메타는 삭제/수정 불가능")
 
         //fixme
         return meta.id!!
@@ -23,7 +24,19 @@ class ReviewMetaServiceImpl(
 
     override fun deleteMeta(id: Long) {
         val meta = reviewMetaRepository.findByIdOrNull(id) ?: throw NotFoundException()
+
+        if (meta.isOpen) throw IllegalArgumentException("open 상태 메타는 삭제/수정 불가능")
         reviewMetaRepository.deleteById(id)
+    }
+
+    override fun open(loginUserId: Long, metaId: Long){
+        val meta = reviewMetaRepository.findByIdOrNull(metaId) ?: throw NotFoundException()
+        meta.open()
+    }
+
+    override fun close(loginUserId: Long, metaId: Long){
+        val meta = reviewMetaRepository.findByIdOrNull(metaId) ?: throw NotFoundException()
+        meta.close()
     }
 
     @Transactional(readOnly = true)
